+++ application/src/main/java/core/application/movies/repositories/comment/CommentRepository.java
@@ -1,10 +1,10 @@
-package core.application.movies.repositories;
+package core.application.movies.repositories.comment;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
-import core.application.movies.models.dto.CommentRespDTO;
+import core.application.movies.models.dto.response.CommentRespDTO;
 import core.application.movies.models.entities.CommentEntity;
 
 /**
