+++ application/src/main/resources/mappers/movies/CommentMapper.xml
@@ -19,7 +19,7 @@
         where movie_id=#{movieId} and user_id=#{userId}
     </select>
 
-    <select id="findByMovieId" resultType="core.application.movies.models.dto.CommentRespDTO">
+    <select id="findByMovieId" resultType="core.application.movies.models.dto.response.CommentRespDTO">
         select c.comment_id, c.content, c.`like`, c.dislike, c.rating, c.movie_id, c.user_id, c.created_at,
             case
                 when l.comment_like_id is not null then true
@@ -43,7 +43,7 @@
         offset #{offset}
     </select>
 
-    <select id="findByMovieIdOnDateDescend" resultType="core.application.movies.models.dto.CommentRespDTO">
+    <select id="findByMovieIdOnDateDescend" resultType="core.application.movies.models.dto.response.CommentRespDTO">
         select c.comment_id, c.content, c.`like`, c.dislike, c.rating, c.movie_id, c.user_id, c.created_at,
             case
                 when l.comment_like_id is not null then true
@@ -68,7 +68,7 @@
         offset #{offset}
     </select>
 
-    <select id="findByMovieIdOnLikeDescend" resultType="core.application.movies.models.dto.CommentRespDTO">
+    <select id="findByMovieIdOnLikeDescend" resultType="core.application.movies.models.dto.response.CommentRespDTO">
         select c.comment_id, c.content, c.`like`, c.dislike, c.rating, c.movie_id, c.user_id, c.created_at,
             case
                 when l.comment_like_id is not null then true
@@ -93,7 +93,7 @@
         offset #{offset}
     </select>
 
-    <select id="findByMovieIdOnDislikeDescend" resultType="core.application.movies.models.dto.CommentRespDTO">
+    <select id="findByMovieIdOnDislikeDescend" resultType="core.application.movies.models.dto.response.CommentRespDTO">
         select c.comment_id, c.content, c.`like`, c.dislike, c.rating, c.movie_id, c.user_id, c.created_at,
             case
                 when l.comment_like_id is not null then true
